brew leaves
In brew terms, leaves are packages that have nothing depending on them,
which is a good approximation for "the packages you really want".
Non-leaves are typically packages that have been auto-installed as
dependencies.

With this change, we only track the version of leaves, which should make
it a bit easier to compare across machines and make informed decisions.