zsh: pick a commit
When working with a git repo, I sometimes work with a single branch
locally, so I can see the result of my combined changes, but want to
make separate PRs for each change individually.

The `pick a b` command will checkout `origin/main`, cherry-pick the `a`
commit on top of it, and push the result to the `origin/b` branch. If
everything went well, it then moves us back to where we were.

There is currently no error handling: if something goes wrong, the user
is left wherever somthing went wrong and has to clean up themselves.
(Typically: `git cherry-pick --abort` followed by `git checkout
wherever-you-were`.)