tmux: explicitly set escape-time
`escape-time` is the number of milliseconds tmux will wait after
registering the ESCkey to see whether it's a standalone escape or part
of a combination of keys.

I frequently use the ESC-arrow combinations, and my reflexes are not
fast enough for the new default value. `500` might be a tad too long,
though; I'll see how things go with more sue.