nix: default nixpkgs (#56)
This adds the option of having a "default" nixpkgs version for a given
user. This somewhat reduces the burden of having to explicitly call
`update-nix` in multiple projects, and reducs the total disk space used
by Nix as it improves the chances that multiple projects will be using
the same versions.

I am not entirely sure how this will interact with git, which seems to
tend to replace files rather than update them in place.