init-nix: fix typo
Introduced in #56; note that `]` is technically not syntax in Bash, so
without the space this is equivalent to:

```bash
if [ "${1:-}" = "--default]" && [ -f "$DEFAULT_SRC" ]; then
```

I don't know what happened with #56, but this is the third PR I'm making
to fix it, and there's another one coming right after. I guess I was a
bit tired that day. :thinking: