zsh/pick: use local branch name
Instead of generating a random name, `pick` will now use the local
branch name as a prefix and append an increasing number, so hopefully
branches are a bit more understandable.

Who cares about branch names, though? The main reason for doing this
change is to make it easier in the future to write a script that would
rebase a collection of linked open PRs.